requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aioice==0.9.0
    # via aiortc
aiortc==1.9.0
    # via streamlit-webrtc
albucore==0.0.11
    # via albumentations
albumentations==1.4.9
    # via insightface
altair==5.3.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
async-timeout==4.0.3
    # via redis
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
av==12.1.0
    # via
    #   -r requirements.in
    #   aiortc
bcrypt==4.1.3
    # via streamlit-authenticator
blinker==1.8.2
    # via streamlit
cachetools==5.3.3
    # via streamlit
certifi==2024.6.2
    # via requests
cffi==1.16.0
    # via
    #   aiortc
    #   cryptography
    #   pylibsrtp
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   tqdm
contourpy==1.2.1
    # via matplotlib
cryptography==42.0.8
    # via
    #   aiortc
    #   pyopenssl
cycler==0.12.1
    # via matplotlib
cython==3.0.10
    # via insightface
dnspython==2.6.1
    # via aioice
easydict==1.13
    # via insightface
extra-streamlit-components==0.1.71
    # via streamlit-authenticator
fonttools==4.53.0
    # via matplotlib
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
google-crc32c==1.5.0
    # via aiortc
idna==3.7
    # via requests
ifaddr==0.2.0
    # via aioice
imageio==2.34.1
    # via scikit-image
insightface==0.7.3
    # via -r requirements.in
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
joblib==1.4.2
    # via scikit-learn
jsonschema==4.22.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
kiwisolver==1.4.5
    # via matplotlib
lazy-loader==0.4
    # via scikit-image
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
matplotlib==3.9.0
    # via insightface
mdurl==0.1.2
    # via markdown-it-py
networkx==3.3
    # via scikit-image
numpy==1.26.4
    # via
    #   -r requirements.in
    #   albucore
    #   albumentations
    #   altair
    #   contourpy
    #   imageio
    #   insightface
    #   matplotlib
    #   onnx
    #   opencv-contrib-python
    #   opencv-python
    #   opencv-python-headless
    #   pandas
    #   pyarrow
    #   pydeck
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   streamlit
    #   tifffile
onnx==1.16.1
    # via insightface
onnxruntime

opencv-contrib-python==4.10.0.84
    # via -r requirements.in
opencv-python==4.10.0.84
    # via -r requirements.in
opencv-python-headless==4.10.0.84
    # via
    #   -r requirements.in
    #   albucore
    #   albumentations
packaging==24.1
    # via
    #   altair
    #   lazy-loader
    #   matplotlib
    #   scikit-image
    #   streamlit
    #   streamlit-webrtc
pandas==2.2.2
    # via
    #   -r requirements.in
    #   altair
    #   streamlit
pillow==10.3.0
    # via
    #   -r requirements.in
    #   imageio
    #   insightface
    #   matplotlib
    #   scikit-image
    #   streamlit
prettytable==3.10.0
    # via insightface
protobuf==4.25.3
    # via
    #   onnx
    #   streamlit
pyarrow==16.1.0
    # via streamlit
pycparser==2.22
    # via cffi
pydantic==2.7.4
    # via albumentations
pydantic-core==2.18.4
    # via pydantic
pydeck==0.9.1
    # via streamlit
pyee==11.1.0
    # via aiortc
pygments==2.18.0
    # via rich
pyjwt==2.8.0
    # via streamlit-authenticator
pylibsrtp==0.10.0
    # via aiortc
pyopenssl==24.1.0
    # via aiortc
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   albumentations
    #   streamlit-authenticator
redis==5.0.6
    # via -r requirements.in
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   insightface
    #   streamlit
rich==13.7.1
    # via streamlit
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
scikit-image==0.24.0
    # via
    #   albumentations
    #   insightface
scikit-learn==1.5.0
    # via
    #   -r requirements.in
    #   albumentations
    #   insightface
scipy==1.13.1
    # via
    #   albumentations
    #   insightface
    #   scikit-image
    #   scikit-learn
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
streamlit==1.35.0
    # via
    #   -r requirements.in
    #   extra-streamlit-components
    #   streamlit-authenticator
    #   streamlit-webrtc
streamlit-authenticator==0.3.2
    # via -r requirements.in
streamlit-webrtc==0.47.7
    # via -r requirements.in
tenacity==8.4.1
    # via streamlit
threadpoolctl==3.5.0
    # via scikit-learn
tifffile==2024.6.18
    # via scikit-image
toml==0.10.2
    # via streamlit
tomli==2.0.1
    # via albucore
toolz==0.12.1
    # via altair
tornado==6.4.1
    # via streamlit
tqdm==4.66.4
    # via insightface
typing-extensions==4.12.2
    # via
    #   albumentations
    #   altair
    #   pydantic
    #   pydantic-core
    #   pyee
    #   streamlit
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via requests
watchdog==4.0.1
    # via streamlit
wcwidth==0.2.13
    # via prettytable
dotenv
fast2sms
